+++ packages/frontend/src/app/dashboard/area/detalle/[area_id]/page.tsx
@@ -1,0 +1,12 @@
+// sections
+import AreaDetailView from 'src/sections/area/detalle/view'
+
+// ----------------------------------------------------------------------
+
+export const metadata = {
+  title: 'Area: Detalle',
+}
+
+export default function Page({ params }: { params: { area_id: string } }) {
+  return <AreaDetailView areaId={params.area_id} />
+}
