--- packages/frontend/src/app/dashboard/area/detalle/[id]/page.tsx
@@ -1,12 +1,0 @@
-// sections
-import AreaDetailView from 'src/sections/area/view'
-
-// ----------------------------------------------------------------------
-
-export const metadata = {
-  title: 'Dashboard: Area',
-}
-
-export default function Page() {
-  return <AreaDetailView />
-}
