+++ packages/frontend/src/sections/area/listado/modal-create.tsx
@@ -1,15 +1,7 @@
 'use client'
 
 import React from 'react'
-import {
-  Typography,
-  Button,
-  Modal,
-  Box,
-  TextField,
-  Grid,
-  Backdrop,
-} from '@mui/material'
+import { Typography, Button, Modal, Box, TextField, Grid, Backdrop } from '@mui/material'
 import Iconify from 'src/components/iconify'
 import { useFormik, FormikHelpers } from 'formik'
 import { useRouter } from 'next/navigation'
