+++ packages/frontend/src/sections/area/listado/view.tsx
